refactor(rust): move provider tests into sibling modules
